fixing license tag in package.xml to conform to SPDX
Updating the package.xml license tags to hold valid SPDX qualifiers.
The rationale behind this is, that with
[meta-ros](https://github.com/ros/meta-ros) and
[superflore](https://github.com/ros-infrastructure/superflore)
recipes get generated for the OpenEmbedded Linux build system.
It will fetch the license tag and copy it over into the recipe. If
no valid SPDX identifiers get found, it will issue an error, and a
manual step to fix this is needed.

See the SPDX identifiers [CC0-1.0](https://spdx.org/licenses/CC0-1.0) and
[BSL-1.0](https://spdx.org/licenses/BSL-1.0.html)

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>